Exercises 4.20* & 4.21*

+++ part4/bloglist/controllers/blogs.js
@@ -3,14 +3,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = (request) => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
-
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
@@ -19,9 +11,9 @@ blogRouter.get('/', async (request, response) => {
 blogRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const decodedToken = jwt.decode(getTokenFrom(request), process.env.SECRET)
+  const decodedToken = jwt.decode(request.token, process.env.SECRET)
 
-  if (!decodedToken.id)
+  if (!decodedToken)
     return response.status(401).json({ error: 'token invalid' })
 
   const user = await User.findById(decodedToken.id)
@@ -42,6 +34,19 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.decode(request.token, process.env.SECRET)
+
+  if (!decodedToken)
+    return response.status(401).json({ error: 'token invalid' })
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  console.log(user)
+
+  if (blog.user.toString() !== user._id.toString()) {
+    response.status(401).json({ error: 'Invalid User' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
